Checking in a localization
The only control that needs localization is the SearchBox. The SearchBox
now has a public property called "strings", that will default to English
if no other strings are provided, but can be overriden by the developer.

There is also a public property exposed called usesThreeCharacterSets.
If this property is true, then the toggle will toggle through alphabet
1, alphabet 2 and numbers and symbols. This would be used for ru-ru
(Russian) and any other language with more than one alphabet.