Enable TVJS.MediaPlayer to be used with WinJS
This change enables the MediaPlayer control to be used with WinJS. The
previous version of the MediaPlayer internally used parts of WinJS.

Now the MediaPlayer's internal references to WinJS have been put under
the TVJS namespace so there are no longer namespace collisions.

This change also fixes the casting button (Miracast).

The tests were updated as well so they all pass with this change.